+++ internal/test-utils-core/src/androidMain/kotlin/com/github/panpf/sketch/test/utils/utils.android.kt
@@ -60,7 +60,7 @@ fun ImageRequest.decode(sketch: Sketch): DecodeResult {
     return BitmapFactoryDecoder(
         requestContext = requestContext,
         dataSource = fetchResult.dataSource.asOrThrow()
-    ).decode().getOrThrow()
+    ).decode()
 }
 
 @Suppress("DEPRECATION")