+++ internal/test-utils-core/src/commonMain/kotlin/com/github/panpf/sketch/test/utils/TestErrorDecoder.kt
@@ -27,7 +27,7 @@ class TestErrorDecoder : Decoder {
     override val imageInfo: ImageInfo
         get() = throw UnsupportedOperationException()
 
-    override fun decode(): Result<DecodeResult> {
+    override fun decode(): DecodeResult {
         throw UnsupportedOperationException()
     }