+++ sketch-animated-koralgif/src/main/kotlin/com/github/panpf/sketch/decode/GifDrawableDecoder.kt
@@ -102,7 +102,7 @@ class GifDrawableDecoder(
         }
 
     @WorkerThread
-    override fun decode(): Result<DecodeResult> = kotlin.runCatching {
+    override fun decode(): DecodeResult {
         val imageInfo = imageInfo
         val size = requestContext.size
         val resize = requestContext.computeResize(imageInfo.size)
@@ -149,7 +149,7 @@ class GifDrawableDecoder(
                 }
             }
 
-        DecodeResult(
+        return DecodeResult(
             image = animatableDrawable.asSketchImage(),
             imageInfo = imageInfo,
             dataFrom = dataSource.dataFrom,